+++ plugin/src/main/kotlin/pl/droidsonroids/gradle/ci/UiTestMode.kt
@@ -1,10 +1,5 @@
 package pl.droidsonroids.gradle.ci
 
 enum class UiTestMode {
-    noMinify, minify;
-
-    fun getMinifyEnabled(defaultMinifyEnabled: Boolean) = when (this) {
-        noMinify -> false
-        else -> defaultMinifyEnabled
-    }
+    noMinify, minify
 }