+++ plugin/src/test/kotlin/pl/droidsonroids/gradle/ci/ConnectedUiTestFunctionalTest.kt
@@ -6,11 +6,6 @@ import org.junit.Rule
 import org.junit.Test
 import java.io.File
 
-fun GradleRunner.withJacoco(): GradleRunner {
-    javaClass.classLoader.getResourceAsStream("testkit-gradle.properties").toFile(File(projectDir, "gradle.properties"))
-    return this
-}
-
 class ConnectedUiTestFunctionalTest {
 
     @get:Rule
@@ -28,7 +23,7 @@ class ConnectedUiTestFunctionalTest {
                 .withTestKitDir(temporaryFolder.newFolder())
                 .withArguments(Constants.CONNECTED_UI_TEST_TASK_NAME, "-P${Constants.UI_TEST_MODE_PROPERTY_NAME}=${UiTestMode.noMinify.name}")
                 .withPluginClasspath()
-                .withJacoco()
+                .withJaCoCo()
                 .buildAndFail()
 
         assertThat(result.output).contains("No connected devices")