+++ plugin/src/test/kotlin/pl/droidsonroids/gradle/ci/GradleVersionCheckFunctionalTest.kt
@@ -20,7 +20,7 @@ class GradleVersionCheckFunctionalTest {
                 .withProjectDir(temporaryFolder.root)
                 .withTestKitDir(temporaryFolder.newFolder())
                 .withPluginClasspath()
-                .withJacoco()
+                .withJaCoCo()
                 .buildAndFail()
 
         Assertions.assertThat(result.output).contains("Gradle version $gradleVersion is not supported")