+++ plugin/src/test/kotlin/pl/droidsonroids/gradle/ci/JaCoCoExtensions.kt
@@ -1,0 +1,9 @@
+package pl.droidsonroids.gradle.ci
+
+import org.gradle.testkit.runner.GradleRunner
+import java.io.File
+
+fun GradleRunner.withJaCoCo(): GradleRunner {
+    javaClass.classLoader.getResourceAsStream("testkit-gradle.properties").toFile(File(projectDir, "gradle.properties"))
+    return this
+}