+++ plugin/src/test/kotlin/pl/droidsonroids/gradle/ci/TestableVariantsFunctionalTest.kt
@@ -20,7 +20,7 @@ class TestableVariantsFunctionalTest {
                 .withTestKitDir(temporaryFolder.newFolder())
                 .withProjectDir(temporaryFolder.root)
                 .withPluginClasspath()
-                .withJacoco()
+                .withJaCoCo()
                 .build()
         assertTestableVariants(result, "productionDev", "stagingDebug")
     }
@@ -33,7 +33,7 @@ class TestableVariantsFunctionalTest {
                 .withTestKitDir(temporaryFolder.newFolder())
                 .withProjectDir(temporaryFolder.root)
                 .withPluginClasspath()
-                .withJacoco()
+                .withJaCoCo()
                 .build()
         assertTestableVariants(result, "productionDev", "stagingDebug", "stagingDev", "stagingRelease", "stagingStore")
     }
@@ -46,7 +46,7 @@ class TestableVariantsFunctionalTest {
                 .withTestKitDir(temporaryFolder.newFolder())
                 .withProjectDir(temporaryFolder.root)
                 .withPluginClasspath()
-                .withJacoco()
+                .withJaCoCo()
                 .build()
         assertTestableVariants(result, "productionDev", "stagingDebug", "stagingDev", "stagingRelease", "stagingStore")
     }
@@ -59,7 +59,7 @@ class TestableVariantsFunctionalTest {
                 .withProjectDir(temporaryFolder.root)
                 .withTestKitDir(temporaryFolder.newFolder())
                 .withPluginClasspath()
-                .withJacoco()
+                .withJaCoCo()
                 .build()
         assertTestableVariants(result, "debug")
     }
@@ -72,7 +72,7 @@ class TestableVariantsFunctionalTest {
                 .withProjectDir(temporaryFolder.root)
                 .withTestKitDir(temporaryFolder.newFolder())
                 .withPluginClasspath()
-                .withJacoco()
+                .withJaCoCo()
                 .build()
         assertTestableVariants(result, "proDebug", "proRelease")
     }
@@ -86,7 +86,7 @@ class TestableVariantsFunctionalTest {
                 .withTestKitDir(temporaryFolder.newFolder())
                 .withArguments("connectedMonkeyJenkinsTest")
                 .withPluginClasspath()
-                .withJacoco()
+                .withJaCoCo()
                 .buildAndFail()
     }
 
@@ -98,7 +98,7 @@ class TestableVariantsFunctionalTest {
                 .withProjectDir(temporaryFolder.root)
                 .withTestKitDir(temporaryFolder.newFolder())
                 .withPluginClasspath()
-                .withJacoco()
+                .withJaCoCo()
                 .buildAndFail()
     }